Filters and models
Added a bandpass filter function to the spectrogram script in order to help reduce some of the excess noise. Still need to test to make sure that the function outputs data that plays well with the mlab.specgram. No reason it shouldn't.

Started working on the CNN architecture and the model pipeline.