Spectrograms and randomality
Updated the spectrogram creation to represent a better frequency window.
Created a loop to pull 100 random samples from each language, as opposed to the first 100, and put them in a training folder. Made this default.
Updated create_imgtxt.py to reflect the new folder path for the randomly selected samples.
Updated remove_spectrograms.sh to remove .jpgs instead of .pngs to reflect the change in image representation.