Success??
Edits to the model_keras.py file got it to run. However, accuracy is < 10%.
    -Reduced image size
    -Reduced batch size

Figured I messed something up with the labels, so I made a new script, model_keras_dirGen.py, using the flow_from_directory function in Keras to create the necessary labels and read in the training images. This new script achieves 99% training accuracy. Going to make a validation and test set to see if this is real life.