Bunch of testing stuff
Made an actual test list containing the files that were NOT included in training and validation.
Made spectrograms for the test set.
Made folder for housing grad_CAMS.
Updated the grad_CAM script to allow for user designation of the file of interest.
    Will update the file to allow for layer selection at some point.
Updated the model itself to use 2x2 strides on the first and second convolutional layers.
    Doing so increased the accuracy on the test files by a marked amount.